Reorganize poly factors methods
