+++ src/galois/_polys/_poly.py
@@ -28,12 +28,6 @@ from ._conversions import (
 SPARSE_VS_DENSE_POLY_FACTOR = 0.00_125  # 1.25% density
 SPARSE_VS_DENSE_POLY_MIN_COEFFS = int(1 / SPARSE_VS_DENSE_POLY_FACTOR)
 
-# Functions that will be monkey-patched in _polys/__init__.py to the actual implementations. This is required because
-# the implementations are in a different module that needs to import Poly. This monkey patching avoids the circular
-# import. If you have a better solution, please open a GitHub issue or pull request.
-def GCD(x: PolyLike, y: PolyLike) -> Poly:
-    raise NotImplementedError
-
 
 @export
 class Poly:
@@ -774,382 +768,6 @@ class Poly:
 
         multiplicities = np.array([_root_multiplicity(self, root) for root in roots])
         return roots, multiplicities
-
-    def square_free_factors(self) -> tuple[list[Poly], list[int]]:
-        r"""
-        Factors the monic polynomial :math:`f(x)` into a product of square-free polynomials.
-
-        Returns:
-            - The list of non-constant, square-free polynomials :math:`h_j(x)` in the factorization.
-            - The list of corresponding multiplicities :math:`j`.
-
-        Raises:
-            ValueError: If :math:`f(x)` is not monic or has degree 0.
-
-        Notes:
-            The Square-Free Factorization algorithm factors :math:`f(x)` into a product of :math:`m` square-free
-            polynomials :math:`h_j(x)` with multiplicity :math:`j`.
-
-            .. math::
-                f(x) = \prod_{j=1}^{m} h_j(x)^j
-
-            Some :math:`h_j(x) = 1`, but those polynomials are not returned by this function.
-
-            A complete polynomial factorization is implemented in :func:`~Poly.factors`.
-
-        References:
-            - Hachenberger, D. and Jungnickel, D. Topics in Galois Fields. Algorithm 6.1.7.
-            - Section 2.1 from https://people.csail.mit.edu/dmoshkov/courses/codes/poly-factorization.pdf
-
-        Examples:
-            Suppose :math:`f(x) = x(x^3 + 2x + 4)(x^2 + 4x + 1)^3` over :math:`\mathrm{GF}(5)`. Each polynomial
-            :math:`x`, :math:`x^3 + 2x + 4`, and :math:`x^2 + 4x + 1` are all irreducible over :math:`\mathrm{GF}(5)`.
-
-            .. ipython:: python
-
-                GF = galois.GF(5)
-                a = galois.Poly([1,0], field=GF); a, a.is_irreducible()
-                b = galois.Poly([1,0,2,4], field=GF); b, b.is_irreducible()
-                c = galois.Poly([1,4,1], field=GF); c, c.is_irreducible()
-                f = a * b * c**3; f
-
-            The square-free factorization is :math:`\{x(x^3 + 2x + 4), x^2 + 4x + 1\}` with multiplicities
-            :math:`\{1, 3\}`.
-
-            .. ipython:: python
-
-                f.square_free_factors()
-                [a*b, c], [1, 3]
-
-        Group:
-            Factorization methods
-
-        Order:
-            51
-        """
-        if not self.degree >= 1:
-            raise ValueError(f"The polynomial must be non-constant, not {self}.")
-        if not self.is_monic:
-            raise ValueError(f"The polynomial must be monic, not {self}.")
-
-        field = self.field
-        p = field.characteristic
-        one = Poly.One(field=field)
-
-        factors_ = []
-        multiplicities = []
-
-        # w is the product (without multiplicity) of all factors of f that have multiplicity not divisible by p
-        d = GCD(self, self.derivative())
-        w = self // d
-
-        # Step 1: Find all factors in w
-        i = 1
-        while w != one:
-            y = GCD(w, d)
-            z = w // y
-            if z != one and i % p != 0:
-                factors_.append(z)
-                multiplicities.append(i)
-            w = y
-            d = d // y
-            i = i + 1
-        # d is now the product (with multiplicity) of the remaining factors of f
-
-        # Step 2: Find all remaining factors (their multiplicities are divisible by p)
-        if d != one:
-            degrees = [degree // p for degree in d.nonzero_degrees]
-            # The inverse Frobenius automorphism of the coefficients
-            coeffs = d.nonzero_coeffs ** (field.characteristic ** (field.degree - 1))
-            delta = Poly.Degrees(degrees, coeffs=coeffs, field=field)  # The p-th root of d(x)
-            f, m = delta.square_free_factors()
-            factors_.extend(f)
-            multiplicities.extend([mi * p for mi in m])
-
-        # Sort the factors in increasing-multiplicity order
-        factors_, multiplicities = zip(*sorted(zip(factors_, multiplicities), key=lambda item: item[1]))
-
-        return list(factors_), list(multiplicities)
-
-    def distinct_degree_factors(self) -> tuple[list[Poly], list[int]]:
-        r"""
-        Factors the monic, square-free polynomial :math:`f(x)` into a product of polynomials whose irreducible factors
-        all have the same degree.
-
-        Returns:
-            - The list of polynomials :math:`f_i(x)` whose irreducible factors all have degree :math:`i`.
-            - The list of corresponding distinct degrees :math:`i`.
-
-        Raises:
-            ValueError: If :math:`f(x)` is not monic, has degree 0, or is not square-free.
-
-        Notes:
-            The Distinct-Degree Factorization algorithm factors a square-free polynomial :math:`f(x)` with degree
-            :math:`d` into a product of :math:`d` polynomials :math:`f_i(x)`, where :math:`f_i(x)` is the product of
-            all irreducible factors of :math:`f(x)` with degree :math:`i`.
-
-            .. math::
-                f(x) = \prod_{i=1}^{d} f_i(x)
-
-            For example, suppose :math:`f(x) = x(x + 1)(x^2 + x + 1)(x^3 + x + 1)(x^3 + x^2 + 1)` over
-            :math:`\mathrm{GF}(2)`, then the distinct-degree factorization is
-
-            .. math::
-                f_1(x) &= x(x + 1) = x^2 + x \\
-                f_2(x) &= x^2 + x + 1 \\
-                f_3(x) &= (x^3 + x + 1)(x^3 + x^2 + 1) = x^6 + x^5 + x^4 + x^3 + x^2 + x + 1 \\
-                f_i(x) &= 1\ \textrm{for}\ i = 4, \dots, 10.
-
-            Some :math:`f_i(x) = 1`, but those polynomials are not returned by this function. In this example,
-            the function returns :math:`\{f_1(x), f_2(x), f_3(x)\}` and :math:`\{1, 2, 3\}`.
-
-            The Distinct-Degree Factorization algorithm is often applied after the Square-Free Factorization algorithm,
-            see :func:`~Poly.square_free_factors`. A complete polynomial factorization is implemented in
-            :func:`~Poly.factors`.
-
-        References:
-            - Hachenberger, D. and Jungnickel, D. Topics in Galois Fields. Algorithm 6.2.2.
-            - Section 2.2 from https://people.csail.mit.edu/dmoshkov/courses/codes/poly-factorization.pdf
-
-        Examples:
-            From the example in the notes, suppose :math:`f(x) = x(x + 1)(x^2 + x + 1)(x^3 + x + 1)(x^3 + x^2 + 1)`
-            over :math:`\mathrm{GF}(2)`.
-
-            .. ipython:: python
-
-                a = galois.Poly([1, 0]); a, a.is_irreducible()
-                b = galois.Poly([1, 1]); b, b.is_irreducible()
-                c = galois.Poly([1, 1, 1]); c, c.is_irreducible()
-                d = galois.Poly([1, 0, 1, 1]); d, d.is_irreducible()
-                e = galois.Poly([1, 1, 0, 1]); e, e.is_irreducible()
-                f = a * b * c * d * e; f
-
-            The distinct-degree factorization is :math:`\{x(x + 1), x^2 + x + 1, (x^3 + x + 1)(x^3 + x^2 + 1)\}`
-            whose irreducible factors have degrees :math:`\{1, 2, 3\}`.
-
-            .. ipython:: python
-
-                f.distinct_degree_factors()
-                [a*b, c, d*e], [1, 2, 3]
-
-        Group:
-            Factorization methods
-
-        Order:
-            51
-        """
-        if not self.degree >= 1:
-            raise ValueError(f"The polynomial must be non-constant, not {self}.")
-        if not self.is_monic:
-            raise ValueError(f"The polynomial must be monic, not {self}.")
-        if not self.is_square_free():
-            raise ValueError(f"The polynomial must be square-free, not {self}.")
-
-        field = self.field
-        q = field.order
-        n = self.degree
-        one = Poly.One(field=field)
-        x = Poly.Identity(field=field)
-
-        factors_ = []
-        degrees = []
-
-        a = self
-        h = x
-
-        l = 1
-        while l <= n // 2 and a != one:
-            h = pow(h, q, a)
-            z = GCD(a, h - x)
-            if z != one:
-                factors_.append(z)
-                degrees.append(l)
-                a = a // z
-                h = h % a
-            l += 1
-
-        if a != one:
-            factors_.append(a)
-            degrees.append(a.degree)
-
-        return factors_, degrees
-
-    def equal_degree_factors(self, degree: int) -> list[Poly]:
-        r"""
-        Factors the monic, square-free polynomial :math:`f(x)` of degree :math:`rd` into a product of :math:`r`
-        irreducible factors with degree :math:`d`.
-
-        Arguments:
-            degree: The degree :math:`d` of each irreducible factor of :math:`f(x)`.
-
-        Returns:
-            The list of :math:`r` irreducible factors :math:`\{g_1(x), \dots, g_r(x)\}` in
-            lexicographically-increasing order.
-
-        Raises:
-            ValueError: If :math:`f(x)` is not monic, has degree 0, or is not square-free.
-
-        Notes:
-            The Equal-Degree Factorization algorithm factors a square-free polynomial :math:`f(x)` with degree
-            :math:`rd` into a product of :math:`r` irreducible polynomials each with degree :math:`d`. This function
-            implements the Cantor-Zassenhaus algorithm, which is probabilistic.
-
-            The Equal-Degree Factorization algorithm is often applied after the Distinct-Degree Factorization algorithm,
-            see :func:`~Poly.distinct_degree_factors`. A complete polynomial factorization is implemented in
-            :func:`~Poly.factors`.
-
-        References:
-            - Section 2.3 from https://people.csail.mit.edu/dmoshkov/courses/codes/poly-factorization.pdf
-            - Section 1 from https://www.csa.iisc.ac.in/~chandan/courses/CNT/notes/lec8.pdf
-
-        Examples:
-            Factor a product of degree-1 irreducible polynomials over :math:`\mathrm{GF}(2)`.
-
-            .. ipython:: python
-
-                a = galois.Poly([1, 0]); a, a.is_irreducible()
-                b = galois.Poly([1, 1]); b, b.is_irreducible()
-                f = a * b; f
-                f.equal_degree_factors(1)
-
-            Factor a product of degree-3 irreducible polynomials over :math:`\mathrm{GF}(5)`.
-
-            .. ipython:: python
-
-                GF = galois.GF(5)
-                a = galois.Poly([1, 0, 2, 1], field=GF); a, a.is_irreducible()
-                b = galois.Poly([1, 4, 4, 4], field=GF); b, b.is_irreducible()
-                f = a * b; f
-                f.equal_degree_factors(3)
-
-        Group:
-            Factorization methods
-
-        Order:
-            51
-        """
-        verify_isinstance(degree, int)
-        if not self.degree >= 1:
-            raise ValueError(f"The polynomial must be non-constant, not {self}.")
-        if not self.is_monic:
-            raise ValueError(f"The polynomial must be monic, not {self}.")
-        if not self.degree % degree == 0:
-            raise ValueError(
-                f"Argument 'degree' must be divide the degree of the polynomial, "
-                f"{degree} does not divide {self.degree}."
-            )
-        if not self.is_square_free():
-            raise ValueError(f"The polynomial must be square-free, not {self}.")
-
-        field = self.field
-        q = field.order
-        r = self.degree // degree
-        one = Poly.One(field)
-
-        factors_ = [self]
-        while len(factors_) < r:
-            h = Poly.Random(degree, field=field)
-            g = GCD(self, h)
-            if g == one:
-                g = pow(h, (q**degree - 1) // 2, self) - one
-            i = 0
-            for u in list(factors_):
-                if u.degree <= degree:
-                    continue
-                d = GCD(g, u)
-                if d not in [one, u]:
-                    factors_.remove(u)
-                    factors_.append(d)
-                    factors_.append(u // d)
-                i += 1
-
-        # Sort the factors in lexicographically-increasing order
-        factors_ = sorted(factors_, key=int)
-
-        return factors_
-
-    def factors(self) -> tuple[list[Poly], list[int]]:
-        r"""
-        Computes the irreducible factors of the non-constant, monic polynomial :math:`f(x)`.
-
-        Returns:
-            - Sorted list of irreducible factors :math:`\{g_1(x), g_2(x), \dots, g_k(x)\}` of :math:`f(x)` sorted in
-              lexicographically-increasing order.
-            - List of corresponding multiplicities :math:`\{e_1, e_2, \dots, e_k\}`.
-
-        Raises:
-            ValueError: If :math:`f(x)` is not monic or has degree 0.
-
-        Notes:
-            This function factors a monic polynomial :math:`f(x)` into its :math:`k` irreducible factors such that
-            :math:`f(x) = g_1(x)^{e_1} g_2(x)^{e_2} \dots g_k(x)^{e_k}`.
-
-            Steps:
-
-            1. Apply the Square-Free Factorization algorithm to factor the monic polynomial into square-free
-               polynomials. See :func:`~Poly.square_free_factors`.
-            2. Apply the Distinct-Degree Factorization algorithm to factor each square-free polynomial into a product
-               of factors with the same degree. See :func:`~Poly.distinct_degree_factors`.
-            3. Apply the Equal-Degree Factorization algorithm to factor the product of factors of equal degree into
-               their irreducible factors. See :func:`~Poly.equal_degree_factors`.
-
-        References:
-            - Hachenberger, D. and Jungnickel, D. Topics in Galois Fields. Algorithm 6.1.7.
-            - Section 2.1 from https://people.csail.mit.edu/dmoshkov/courses/codes/poly-factorization.pdf
-
-        Examples:
-            Generate irreducible polynomials over :math:`\mathrm{GF}(3)`.
-
-            .. ipython:: python
-
-                GF = galois.GF(3)
-                g1 = galois.irreducible_poly(3, 3); g1
-                g2 = galois.irreducible_poly(3, 4); g2
-                g3 = galois.irreducible_poly(3, 5); g3
-
-            Construct a composite polynomial.
-
-            .. ipython:: python
-
-                e1, e2, e3 = 5, 4, 3
-                f = g1**e1 * g2**e2 * g3**e3; f
-
-            Factor the polynomial into its irreducible factors over :math:`\mathrm{GF}(3)`.
-
-            .. ipython:: python
-
-                f.factors()
-
-        Group:
-            Factorization methods
-
-        Order:
-            51
-        """
-        if not self.degree >= 1:
-            raise ValueError(f"The polynomial must be non-constant, not {self}.")
-        if not self.is_monic:
-            raise ValueError(f"The polynomial must be monic, not {self}.")
-
-        factors_, multiplicities = [], []
-
-        # Step 1: Find all the square-free factors
-        sf_factors, sf_multiplicities = self.square_free_factors()
-
-        # Step 2: Find all the factors with distinct degree
-        for sf_factor, sf_multiplicity in zip(sf_factors, sf_multiplicities):
-            df_factors, df_degrees = sf_factor.distinct_degree_factors()
-
-            # Step 3: Find all the irreducible factors with degree d
-            for df_factor, df_degree in zip(df_factors, df_degrees):
-                f = df_factor.equal_degree_factors(df_degree)
-                factors_.extend(f)
-                multiplicities.extend([sf_multiplicity] * len(f))
-
-        # Sort the factors in increasing-multiplicity order
-        factors_, multiplicities = zip(*sorted(zip(factors_, multiplicities), key=lambda item: int(item[0])))
-
-        return list(factors_), list(multiplicities)
 
     def derivative(self, k: int = 1) -> Poly:
         r"""
